FedoraXFCE/bin/speakflow/requirements.txt
openai-whisper
sounddevice
scipy
numpy
pynput
pystray
Pillow
pyperclip
